Fixed jwt, now you cannot authorize using refresh token, token type split from BEARER to ACCESS and RREFRESH, improved faulty jwt logic

+++ src/main/java/com/example/shopberry/auth/jwt/TokenRepository.java
@@ -5,9 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface TokenRepository extends JpaRepository<Token, Integer> {
+public interface TokenRepository extends JpaRepository<Token, Long> {
 
-    List<Token> findByUser_IdAndIsExpiredFalseOrIsRevokedFalse(Integer userId);
+    List<Token> findByUser_IdAndIsExpiredFalseOrIsRevokedFalse(Long userId);
 
     Optional<Token> findByToken(String token);
